Create a control structure for running EM algorithms

Create a control structure for running EM algorithms

em.control(ctrl = NULL, ...)

Arguments

nplot
how often to plot wages of comparaison to model0

ncat
how often to log information

maxiter
maximum number of iterations

model_var
whether to allow for flexible variance (default is TRUE)

est_Amb
TBD

cstr_type
defines the type of constraints on the means for estimation

cstr_val
TBD

tol
tolerance for stopping the EM

tau
posterior likelihood to use instead of computing them using initial parameters

model0
model to compare estimation to when plotting

presp=1e-9
TBD

rel_weight
setting a particular weight for stayers versus movers (default is 1)

est_rho
vector of TRUE/FALSE that states which rho should be estimated

rho_in_diff
whether to estimate rho in level of in differences

dprior
Dirichlet prior for proportions (default 1.01)

nfirms_to_update=10,
number firms to update in probabilistic approach

proba_include
what terms to include in the liklihood for probabilistic approach (default to = c(1,1,1,1))

check_lik
whether to check the likelihood at each updating parameter

stochastic
whether to use stochastic EM instead of straight EM

fixb
when TRUE, imposes fixed interactions in different time periods

fixm
when TRUE, levels are not updated, only variances and proportions

deps=1e-50
TBD

file_backup_prefix
TBD

sd_floor=
floor imposed on all standard deviations (default is 1e-10)

posterior_reg
term added to posterior probablities (this is to deal with numerical issues, default is 1e-9)

textapp
text to show in logging

sdata_subsample
share of the stayers to use in estimation

sdata_subredraw
whether to redraw the subsample of stayers

vdec_sim_size
size to use in simulation

stayer_weight
weight attributed to stayers in joint estimation

est_rep=
number of starting values for EM

est_nbest
number of best starting values to select from using connectedness